+++ site/dat/docs/Changelog.md
@@ -3,6 +3,8 @@
 ## 0.4.2 (next)
  - bump up gatling version to 2.1.7
  - selenium script from requests format
+ - change file search algo to use paths relative to config locations
+ - allow having script files relative to config locations
 
 ## 0.4.1
  - fix Locust.io search paths